Agora lista é paginado e ordenado de acordo com os parametros recebidos

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -7,8 +7,11 @@ import br.com.alura.forum.controller.form.TopicoForm;
 import br.com.alura.forum.model.Topico;
 import br.com.alura.forum.repository.CursoRepository;
 import br.com.alura.forum.repository.TopicoRepository;
-import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
@@ -31,20 +34,21 @@ public class TopicosController {
 //    @GetMapping("/topicos/{nomeCurso}")
 //    public List<TopicoDto> lista(@PathVariable(value = "nomeCurso", required = false) final String nomeCurso){
 
-//    http://localhost:8080/topico?nomeCurso=abc
-//    @GetMapping()
-//    public List<TopicoDto> lista(String nomeCurso){}
-
-    @GetMapping({"/lista/{nomeCurso}", "/lista"})
-    public List<TopicoDto> lista(@PathVariable(value = "nomeCurso", required = false) Optional<String> nomeCurso){
-        System.out.println(nomeCurso);
+    //    http://localhost:8080/topico?nomeCurso=abc
+//    @GetMapping({"/lista/{nomeCurso}", "/lista"})
+//    public List<TopicoDto> lista(@PathVariable(value = "nomeCurso", required = false) Optional<String> nomeCurso){
+    @GetMapping()
+    public Page<TopicoDto> lista(@RequestParam(required = false) String nomeCurso,
+                @RequestParam int pagina, @RequestParam int quantidade, @RequestParam() String campoOrdenacao) {
 //        Topico topico = new Topico("Dúvida", "Dúvida com Spring", new Curso("Spring","Programação"));
 //        return TopicoDto.converter(Arrays.asList(topico, topico));
 
-        if (nomeCurso.isEmpty()) {
-            return TopicoDto.converter(topicoRepository.findAll());
+        Pageable paginacao = PageRequest.of(pagina, quantidade, Sort.Direction.DESC, campoOrdenacao);
+
+        if (nomeCurso == null) {
+            return TopicoDto.converter(topicoRepository.findAll(paginacao));
         } else {
-            return TopicoDto.converter(topicoRepository.findByCurso_Nome(nomeCurso.get()));
+            return TopicoDto.converter(topicoRepository.findByCurso_Nome(nomeCurso, paginacao));
         }
     }
 
@@ -61,26 +65,40 @@ public class TopicosController {
     }
 
     @GetMapping("/{id}")
-    public DetalhesDoTopicoDto detalhar(@PathVariable Long id) {
-        Topico topico = topicoRepository.getOne(id);
+    public ResponseEntity<DetalhesDoTopicoDto> detalhar(@PathVariable Long id) {
+        Optional<Topico> topico = topicoRepository.findById(id);
 
-        return new DetalhesDoTopicoDto(topico);
+        if (topico.isPresent()) {
+            return ResponseEntity.ok(new DetalhesDoTopicoDto(topico.get()));
+        }
+
+        return ResponseEntity.notFound().build();
     }
 
     @PutMapping("/{id}")
     @Transactional
     public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id,
                                                @RequestBody @Valid AtualizacaoTopicoForm topicoForm) {
-        Topico topico = topicoForm.topicoComNovosDados(id, topicoRepository);
+        Optional<Topico> topico = topicoRepository.findById(id);
 
-        return ResponseEntity.ok(new TopicoDto(topico));
+        if (topico.isPresent()) {
+            return ResponseEntity.ok(new TopicoDto(topicoForm.atualizar(topico.get())));
+        }
+
+
+        return ResponseEntity.notFound().build();
     }
 
     @DeleteMapping("/{id}")
     @Transactional
     public ResponseEntity<?> remover(@PathVariable Long id) {
-        topicoRepository.deleteById(id);
+        Optional<Topico> topico = topicoRepository.findById(id);
 
-        return ResponseEntity.ok().build();
+        if (topico.isPresent()) {
+            topicoRepository.deleteById(id);
+            return ResponseEntity.ok().build();
+        }
+
+        return ResponseEntity.notFound().build();
     }
 }
